requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.23
pydantic-settings==2.3.4
python-dotenv==1.0.1
alipay-sdk-python==3.3.946
python-multipart==0.0.9
